requirements.txt
# SyntaAI SAP Security MCP Server - OData Edition
# Zero native dependencies - pure Python + HTTP

# MCP SDK
mcp>=1.0.0

# HTTP client (async, modern, fast)
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0
